pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "masis"
version = "1.0.0"
description = "Multi-Agent Strategic Intelligence System - A Chief of Staff AI that reasons, validates, and justifies outcomes"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-community>=0.3.0",
    "langchain-core>=0.3.0",
    "chromadb>=0.5.0",
    "pydantic>=2.0",
    "tiktoken>=0.7.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
eval = ["ragas>=0.1.0", "datasets>=2.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]

[tool.setuptools.packages.find]
include = ["masis*"]

[project.scripts]
masis = "masis.cli:main"
